Rework flow of WP app
Rework the flow of the WP app to more closely align to the feel of the
OS. Starts at the panorama page instead of the scan page first. Also
prevents object-flickering weirdness (i.e. loading the scan page first,
then quickly changing to the waiting pano). Combine the scan page and
manual code entry page into a pivot page. Unfortunately, the QRScanner
does not function if placed inside a pano page, so moving this out to
another page solved this problem.